requirements.txt
google-auth
google-auth-oauthlib
google-analytics-data
supabase
python-dotenv
schedule